Drop the KPopupMenu activated signal dependency.
The KPopupMenu activated signal introduces introduces a extra slot
(slotContextMenuActivated) between the signal emitter (the menu item) and the
slot that performs the real work. By using the functionality provided by
KAction we can directly connect to the end slot.

This is the first of a series of commits that will refactor the KCheckGmail
architecture.