Move functionality away from KCheckGmailTray.
KCheckGmailTray was indeed the core class. It owned every functionality
relevant class in the project. KCheckGmailCore is now the core class. It
also owns the DCOP interface.

Add a JSProtocol (JavaScriptProtocol) class. JSProtocol owns (encapsulates)
GMailParser and GMail.

Change the way mail count updates are handled.The system tray must be
isolated from the internals of a mail count update.It should only receive
the updates and display them.

The Gmail (object) count update signals are now connected to the mJSP
object. mTray object no longer asks for computation of new and unread
messages. The loginDone signal is now connected to the mJSP object. The
mJSP then reemites a loginDone signal with the info to be displayedin the
systemtray. This way we can reset the counters if the connection is lost.

MailCounter contains the current count (both new and unread) and also the
previous.

KCheckGmailTray slotMailCountChanged and slotmailArrived are independent
now. slotMailCountChanged receives the number of unread messages

updateThreadMenu: No longer computes the Threads menu entries.