Show the tray icon earlier.
The tray icon only showed up after the setPixmapAuth() call
inside KCheckGmailTray::slotLoginStart call. Why? The pixmap wasn't set.

One effect was the missing icon in the the Welcome dialog screenshot
(first time user).
Another was the lack of the icon, altough the contextmenu is already
working, until the slotLoginStart is called.

Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>