Port notifications
Ported notifications to the new API.
Kopete KNotification dropped. Using KDE4 KNotification instead.
The config file is now kcheckgmail.notifyrc. The file is quite similar to
the old eventsrc. See kdelibs/knotify/PORTING.html for the details.

KNotification::event() takes an QWidget (0L by default) and
KCheckGmailTray is longer an QWidget descendent. So the parent widget for
the KNotification::event() call is 0L.
TODO: Check how it works and find a fix, if needed.

Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>